"AU_LINUX_ANDROID_LA.BF64.1.2.1_RB2.05.00.02.081.002" based on quic/aosp/LA.BF64.1.2.1_rb2

Conflicts:
	core/java/android/os/Environment.java
	telecomm/java/android/telecom/ParcelableConference.java

+++ packages/Keyguard/src/com/android/keyguard/EmergencyButton.java
@@ -46,7 +46,7 @@ public class EmergencyButton extends Button {
     KeyguardUpdateMonitorCallback mInfoCallback = new KeyguardUpdateMonitorCallback() {
 
         @Override
-        public void onSimStateChanged(long subId, State simState) {
+        public void onSimStateChanged(int subId, State simState) {
             int phoneState = KeyguardUpdateMonitor.getInstance(mContext).getPhoneState();
             updateEmergencyCallButton(phoneState);
         }
