+++ packages/SystemUI/src/com/android/systemui/statusbar/policy/MSimNetworkControllerImpl.java
@@ -43,7 +43,6 @@ import android.telephony.PhoneStateListener;
 import android.telephony.ServiceState;
 import android.telephony.SignalStrength;
 import android.telephony.SubscriptionManager;
-import android.telephony.SubInfoRecord;
 import android.telephony.TelephonyManager;
 import android.text.TextUtils;
 import android.util.Slog;
@@ -107,7 +106,7 @@ public class MSimNetworkControllerImpl extends NetworkControllerImpl {
     int mPhoneCount = 0;
     int PHONE_ID1 = PhoneConstants.PHONE_ID1;
     int PHONE_ID2 = PhoneConstants.PHONE_ID2;
-    private HashMap<Long, Integer> mSubIdPhoneIdMap;
+    private HashMap<Integer, Integer> mSubIdPhoneIdMap;
     ArrayList<MSimSignalCluster> mSimSignalClusters = new ArrayList<MSimSignalCluster>();
     ArrayList<TextView> mSubsLabelViews = new ArrayList<TextView>();
 
@@ -230,14 +229,14 @@ public class MSimNetworkControllerImpl extends NetworkControllerImpl {
         //List<SubInfoRecord> subInfoList = SubscriptionManager.getActivatedSubInfoList(context);
         //if (subInfoList != null) {
             //int subCount = subInfoList.size();
-            mSubIdPhoneIdMap = new HashMap<Long, Integer>();
+            mSubIdPhoneIdMap = new HashMap<Integer, Integer>();
             mPhoneCount = TelephonyManager.getDefault().getPhoneCount();
              Slog.d(TAG, "registerPhoneStateListener: " + mPhoneCount);
             mMSimPhoneStateListener = new PhoneStateListener[mPhoneCount];
             for (int i=0; i < mPhoneCount; i++) {
-                long[] subIdtemp = SubscriptionManager.getSubId(i);
+                int[] subIdtemp = SubscriptionManager.getSubId(i);
                 if (subIdtemp != null) {
-                    long subId = subIdtemp[0];
+                    int subId = subIdtemp[0];
                     Slog.d(TAG, "registerPhoneStateListener subId: "+ subId);
                     Slog.d(TAG, "registerPhoneStateListener slotId: "+ i);
                     //if (subInfoList.get(i).mSubId >= 0) {
@@ -270,7 +269,7 @@ public class MSimNetworkControllerImpl extends NetworkControllerImpl {
     }
 
 
-    private int getPhoneId(long subId) {
+    private int getPhoneId(int subId) {
         int phoneId;
         phoneId = SubscriptionManager.getPhoneId(subId);
         Slog.d(TAG, "getPhoneId phoneId: " +phoneId);
@@ -361,7 +360,7 @@ public class MSimNetworkControllerImpl extends NetworkControllerImpl {
                 refreshViews(sub);
             }
         } else if (action.equals(TelephonyIntents.SPN_STRINGS_UPDATED_ACTION)) {
-            final long subId = intent.getLongExtra(PhoneConstants.SUBSCRIPTION_KEY, 0);
+            final int subId = intent.getIntExtra(PhoneConstants.SUBSCRIPTION_KEY, 0);
             Slog.d(TAG, "Received SPN update on subId :" + subId);
             Integer phoneId = getPhoneId(subId);
             Slog.d(TAG, "Received SPN update on phoneId :" + phoneId);
@@ -505,7 +504,7 @@ public class MSimNetworkControllerImpl extends NetworkControllerImpl {
 
     // ===== Telephony ==============================================================
 
-    private PhoneStateListener getPhoneStateListener(long subId, int slotId) {
+    private PhoneStateListener getPhoneStateListener(int subId, int slotId) {
         PhoneStateListener mMSimPhoneStateListener = new PhoneStateListener(subId) {
             @Override
             public void onSignalStrengthsChanged(SignalStrength signalStrength) {
